feat: add ability to limit number of records
This allows callers to cap the number of records in a given run.

This would have been trivial had we picked an API that didn't pass
around channels, but alas, mistakes were made. As is, we need to read
off one channel, and write to another. Shutdown is done from the write
end.